move coulom matrix to structural describer
